step defintion for default sort moved to common steps, getAuditLog method moved to commonitemspage, removed UI label validation step from audit history and moved the validation to assertOnAuditHistoryReviewBodyPage
