+++ src/resources/test_data/iras/reviewResearch/userAdministration/manageReviewBodies/audit_history_review_body_page_data.json
@@ -4,6 +4,11 @@
     "next_button": "Next   page",
     "first_row_index": "0",
     "system_admin_email_text": "futureiras.automation@hra.nhs.uk",
+    "Column_Header_Labels": {
+      "date_header_label": " Date",
+      "event_description_label": "Event description",
+      "sys_admin_label": "System administrator"
+    },
     "Create_Review_Body": {
       "event_description_text": " was created"
     },
@@ -35,10 +40,5 @@
       "description_index": "2",
       "country_index": "3"
     }
-  },
-  "Review_Body_Audit_History": {
-    "date_label": "Date",
-    "event_description_label": "Event description",
-    "system_admin_label": "System administrator"
   }
 }
